Declare function prototypes for nullary functions
Prior to C23, a function declaration with empty parentheses (like
`void foo();`) declares a function with an unspecified number of
parameters.  Starting with C23, `void foo();` is equivalent to `void
foo(void);`, which declares a function that has no parameters.
Explicitly add `void` for nullary functions so that pre-C23 compilers
will detect the same bugs that C23 compilers detect.